+++ modules/sbr/unit-test-coverage/test_cfe_sbr_map_direct.c
@@ -99,7 +99,7 @@ void Test_SBR_Map_Direct(void)
             count++;
         }
     }
-    UtPrintf("Valid route id's encountered in performance loop: %u", count);
+    UtPrintf("Valid route id's encountered in performance loop: %u", (unsigned int)count);
 }
 
 /* Main unit test routine */
